feat: rest of the features
